+++ packages/core/src/graph/adjacency.ts
@@ -30,6 +30,7 @@ export function getAdjacencyMap<
     serializeEvent,
     serializeState,
     events: getEvents,
+    filterEvents,
     limit,
     fromState: customFromState,
     stopWhen
@@ -87,6 +88,10 @@ export function getAdjacencyMap<
       typeof getEvents === 'function' ? getEvents(state) : getEvents;
 
     for (const nextEvent of events) {
+      if (filterEvents && !filterEvents(state, nextEvent)) {
+        continue;
+      }
+
       const nextSnapshot = transition(state, nextEvent, actorScope);
 
       adj[serializedState].transitions[