+++ packages/core/src/graph/graph.ts
@@ -201,6 +201,7 @@ export function resolveTraversalOptions<TLogic extends AnyActorLogic>(
     serializeState,
     serializeEvent,
     events: [],
+    filterEvents: undefined,
     limit: Infinity,
     fromState: undefined,
     toState: undefined,